Extract API types to API module
This makes it easier to reuse them between modules.
